refactor: RegisterController와 RegisterViewController를 어노테이션 기반으로 리팩토링
